gas-oracle: fix type usage
Convert the average block gas limit to a `uint64` in the
codebase instead of it being used as a `float64`.

+++ go/gas-oracle/gasprices/gas_price_updater.go
@@ -16,7 +16,7 @@ type GasPriceUpdater struct {
 	mu                     *sync.RWMutex
 	gasPricer              *GasPricer
 	epochStartBlockNumber  uint64
-	averageBlockGasLimit   float64
+	averageBlockGasLimit   uint64
 	epochLengthSeconds     uint64
 	getLatestBlockNumberFn GetLatestBlockNumberFn
 	getGasUsedByBlockFn    GetGasUsedByBlockFn
@@ -26,7 +26,7 @@ type GasPriceUpdater struct {
 func NewGasPriceUpdater(
 	gasPricer *GasPricer,
 	epochStartBlockNumber uint64,
-	averageBlockGasLimit float64,
+	averageBlockGasLimit uint64,
 	epochLengthSeconds uint64,
 	getLatestBlockNumberFn GetLatestBlockNumberFn,
 	getGasUsedByBlockFn GetGasUsedByBlockFn,
